feat(enfermeria): implement responsive nurse module for triage, rapid dosing and patient transfers

+++ src/SistemaSatHospitalario.Core.Domain/Entities/Admision/DetalleServicioCuenta.cs
@@ -10,7 +10,7 @@ namespace SistemaSatHospitalario.Core.Domain.Entities.Admision
         public string Descripcion { get; private set; }
         public decimal Precio { get; private set; }
         public decimal Honorario { get; private set; }
-        public int Cantidad { get; private set; }
+        public decimal Cantidad { get; private set; }
         public string TipoServicio { get; private set; } // Medico, RX, Laboratorio, Insumo
         public string UsuarioCarga { get; private set; }
         public DateTime FechaCarga { get; private set; }
@@ -27,7 +27,7 @@ namespace SistemaSatHospitalario.Core.Domain.Entities.Admision
 
         protected DetalleServicioCuenta() { }
 
-        public DetalleServicioCuenta(Guid cuentaServicioId, Guid servicioId, string descripcion, decimal precio, decimal honorario, int cantidad, string tipoServicio, string usuarioCarga, string? legacyMappingId = null)
+        public DetalleServicioCuenta(Guid cuentaServicioId, Guid servicioId, string descripcion, decimal precio, decimal honorario, decimal cantidad, string tipoServicio, string usuarioCarga, string? legacyMappingId = null)
         {
             Id = Guid.NewGuid();
             CuentaServicioId = cuentaServicioId;
